Fixed some admin GUI stuff Added caching PlayURLs Added the ability to delete old files

+++ lua/arclib/shared/file_util.lua
@@ -7,12 +7,32 @@ end
 
 
 function ARCLib.DeleteAll(dir) --Exactly like doing "rm -rf [dir]" on linux
+	if file.IsDir(dir) then
+		local files, directories = file.Find( dir.."/*", "DATA" )
+		for k,v in pairs(files) do
+			file.Delete(dir.."/"..v)
+		end
+		for k,v in pairs(directories) do
+			ARCLib.DeleteAll(dir.."/"..v)
+		end
+		file.Delete(dir)
+	end
+end
+
+function ARCLib.DeleteOldFiles(time,dir,recur,delfol)
+	local dtime = os.time()-time
 	local files, directories = file.Find( dir.."/*", "DATA" )
-	for k,v in pairs(files) do
-		file.Delete(dir.."/"..v)
+	for i=1,#files do
+		if file.Time( dir.."/"..files[i], "DATA") < dtime then
+			file.Delete(dir.."/"..files[i])
+		end
 	end
-	for k,v in pairs(directories) do
-		ARCLib.DeleteAll(dir.."/"..v)
+	if recur then
+		for i=1,#directories do
+			ARCLib.DeleteOldFiles(time,dir.."/"..directories[i],true,delfol)
+		end
+		if delfol then
+			file.Delete(dir)
+		end
 	end
-	file.Delete(dir)
 end
